Teacher Assistance/requirements.txt
Flask==2.3.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==0.19.0
gunicorn==20.1.0
Werkzeug==3.1.3
psycopg2-binary==2.9.9
stripe==7.11.0
PyJWT==2.8.0
SQLAlchemy>=2.0.43
pymysql==1.1.0
mysqlclient==2.2.0
flask-mail==0.9.1
mysql-connector-python==8.1.0
itsdangerous==2.1.2
requests==2.31.0
youtube-transcript-api>=1.0.0
python-json-logger==2.0.7
anthropic>=0.5.0
cryptography==41.0.7
flask-jwt-extended==4.5.2
cohere<5.0.0,>=4.0.0
python-docx==1.0.1
reportlab==4.0.8
pdfkit==1.0.0
google-auth>=2.0.0
google-auth-oauthlib>=0.4.6
google-auth-httplib2>=0.1.0
nltk==3.8.1
matplotlib==3.8.0
PyMySQL>=1.0.0
numpy==1.24.3
plotly==5.17.0
pandas==2.0.3
kaleido==0.2.1
Pillow==10.0.1
boto3==1.34.144
Flask-Limiter==3.5.0
google-api-python-client==2.100.0
twilio==8.10.0